@@ -31,7 +31,7 @@ using namespace Eigen;
 
 typedef message_filters::sync_policies::ApproximateTime<sensor_msgs::Image, geometry_msgs::TransformStamped> approx_policy;
 
-int *depth_hostptr;
+int* depth_hostptr;
 cv::Mat depth_mat;
 
 int width, height;
@@ -64,7 +64,7 @@ vector<PoseInfo> gt_pose_vect;
 
 void render_currentpose();
 
-bool read_pose(fstream &file)
+bool read_pose(fstream& file)
 {
   int count = 0;
   bool good = true;
@@ -104,7 +104,7 @@ bool read_pose(fstream &file)
 
 vector<cv::Point3f> pts_3;
 vector<cv::Point2f> pts_2;
-void imageBackFunc(int event, int x, int y, int flags, void *userdata)
+void imageBackFunc(int event, int x, int y, int flags, void* userdata)
 {
   if (event == EVENT_LBUTTONDOWN)
   {
@@ -112,7 +112,7 @@ void imageBackFunc(int event, int x, int y, int flags, void *userdata)
     pts_2.push_back(cv::Point2f(x, y));
   }
 }
-void depthBackFunc(int event, int x, int y, int flags, void *userdata)
+void depthBackFunc(int event, int x, int y, int flags, void* userdata)
 {
   if (event == EVENT_LBUTTONDOWN)
   {
@@ -159,13 +159,10 @@ void solve_pnp()
   pnp_result(2, 3) = t.at<double>(2, 0);
 
   vicon2leica = pnp_result.inverse();
-  cout << "translation : " << endl
-       << pnp_result << endl;
+  cout << "translation : " << endl << pnp_result << endl;
 }
 
-void image_pose_callback(
-    const sensor_msgs::ImageConstPtr &image_input,
-    const geometry_msgs::TransformStampedConstPtr &pose_input)
+void image_pose_callback(const sensor_msgs::ImageConstPtr& image_input, const geometry_msgs::TransformStampedConstPtr& pose_input)
 {
   // time diff
   double time_diff = fabs(image_input->header.stamp.toSec() - pose_input->header.stamp.toSec()) * 1000.0;
@@ -275,7 +272,7 @@ void render_currentpose()
   cv::imshow("depth_image", adjMap);
 }
 
-int main(int argc, char **argv)
+int main(int argc, char** argv)
 {
   ros::init(argc, argv, "cloud_banchmark");
   ros::NodeHandle nh("~");
@@ -290,10 +287,7 @@ int main(int argc, char **argv)
   depthrender.set_para(fx, fy, cx, cy, width, height);
 
   cv_K = (cv::Mat_<float>(3, 3) << fx, 0.0f, cx, 0.0f, fy, cy, 0.0f, 0.0f, 1.0f);
-  if (nh.hasParam("cam_k1") &&
-      nh.hasParam("cam_k2") &&
-      nh.hasParam("cam_r1") &&
-      nh.hasParam("cam_r2"))
+  if (nh.hasParam("cam_k1") && nh.hasParam("cam_k2") && nh.hasParam("cam_r1") && nh.hasParam("cam_r2"))
   {
     float k1, k2, r1, r2;
     nh.getParam("cam_k1", k1);
@@ -301,14 +295,7 @@ int main(int argc, char **argv)
     nh.getParam("cam_r1", r1);
     nh.getParam("cam_r2", r2);
     cv_D = (cv::Mat_<float>(1, 4) << k1, k2, r1, r2);
-    cv::initUndistortRectifyMap(
-        cv_K,
-        cv_D,
-        cv::Mat_<double>::eye(3, 3),
-        cv_K,
-        cv::Size(width, height),
-        CV_16SC2,
-        undist_map1, undist_map2);
+    cv::initUndistortRectifyMap(cv_K, cv_D, cv::Mat_<double>::eye(3, 3), cv_K, cv::Size(width, height), CV_16SC2, undist_map1, undist_map2);
     is_distorted = true;
   }
   if (is_distorted)
@@ -316,14 +303,10 @@ int main(int argc, char **argv)
   else
     printf("do not need to rectify.\n");
 
-  vicon2body << 0.33638, -0.01749, 0.94156, 0.06901,
-      -0.02078, -0.99972, -0.01114, -0.02781,
-      0.94150, -0.01582, -0.33665, -0.12395,
-      0.0, 0.0, 0.0, 1.0;
-  cam02body << 0.0148655429818, -0.999880929698, 0.00414029679422, -0.0216401454975,
-      0.999557249008, 0.0149672133247, 0.025715529948, -0.064676986768,
-      -0.0257744366974, 0.00375618835797, 0.999660727178, 0.00981073058949,
-      0.0, 0.0, 0.0, 1.0;
+  vicon2body << 0.33638, -0.01749, 0.94156, 0.06901, -0.02078, -0.99972, -0.01114, -0.02781, 0.94150, -0.01582, -0.33665, -0.12395, 0.0, 0.0, 0.0,
+      1.0;
+  cam02body << 0.0148655429818, -0.999880929698, 0.00414029679422, -0.0216401454975, 0.999557249008, 0.0149672133247, 0.025715529948, -0.064676986768,
+      -0.0257744366974, 0.00375618835797, 0.999660727178, 0.00981073058949, 0.0, 0.0, 0.0, 1.0;
   cam2world = Matrix4d::Identity();
 
   string cloud_path;
@@ -350,7 +333,7 @@ int main(int argc, char **argv)
 
   // pass cloud_data to depth render
   depthrender.set_data(cloud_data);
-  depth_hostptr = (int *)malloc(width * height * sizeof(int));
+  depth_hostptr = (int*)malloc(width * height * sizeof(int));
 
   message_filters::Subscriber<sensor_msgs::Image> image_sub(nh, "/cam0/image_raw", 30);
   message_filters::Subscriber<geometry_msgs::TransformStamped> pose_sub(nh, "/vicon/firefly_sbx/firefly_sbx", 30);
