@@ -246,27 +246,27 @@ inline __host__ __device__ uint4 make_uint4(int4 a)
 // negate
 ////////////////////////////////////////////////////////////////////////////////
 
-inline __host__ __device__ float2 operator-(float2 &a)
+inline __host__ __device__ float2 operator-(float2& a)
 {
   return make_float2(-a.x, -a.y);
 }
-inline __host__ __device__ int2 operator-(int2 &a)
+inline __host__ __device__ int2 operator-(int2& a)
 {
   return make_int2(-a.x, -a.y);
 }
-inline __host__ __device__ float3 operator-(float3 &a)
+inline __host__ __device__ float3 operator-(float3& a)
 {
   return make_float3(-a.x, -a.y, -a.z);
 }
-inline __host__ __device__ int3 operator-(int3 &a)
+inline __host__ __device__ int3 operator-(int3& a)
 {
   return make_int3(-a.x, -a.y, -a.z);
 }
-inline __host__ __device__ float4 operator-(float4 &a)
+inline __host__ __device__ float4 operator-(float4& a)
 {
   return make_float4(-a.x, -a.y, -a.z, -a.w);
 }
-inline __host__ __device__ int4 operator-(int4 &a)
+inline __host__ __device__ int4 operator-(int4& a)
 {
   return make_int4(-a.x, -a.y, -a.z, -a.w);
 }
@@ -279,7 +279,7 @@ inline __host__ __device__ float2 operator+(float2 a, float2 b)
 {
   return make_float2(a.x + b.x, a.y + b.y);
 }
-inline __host__ __device__ void operator+=(float2 &a, float2 b)
+inline __host__ __device__ void operator+=(float2& a, float2 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -292,7 +292,7 @@ inline __host__ __device__ float2 operator+(float b, float2 a)
 {
   return make_float2(a.x + b, a.y + b);
 }
-inline __host__ __device__ void operator+=(float2 &a, float b)
+inline __host__ __device__ void operator+=(float2& a, float b)
 {
   a.x += b;
   a.y += b;
@@ -302,7 +302,7 @@ inline __host__ __device__ int2 operator+(int2 a, int2 b)
 {
   return make_int2(a.x + b.x, a.y + b.y);
 }
-inline __host__ __device__ void operator+=(int2 &a, int2 b)
+inline __host__ __device__ void operator+=(int2& a, int2 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -315,7 +315,7 @@ inline __host__ __device__ int2 operator+(int b, int2 a)
 {
   return make_int2(a.x + b, a.y + b);
 }
-inline __host__ __device__ void operator+=(int2 &a, int b)
+inline __host__ __device__ void operator+=(int2& a, int b)
 {
   a.x += b;
   a.y += b;
@@ -325,7 +325,7 @@ inline __host__ __device__ uint2 operator+(uint2 a, uint2 b)
 {
   return make_uint2(a.x + b.x, a.y + b.y);
 }
-inline __host__ __device__ void operator+=(uint2 &a, uint2 b)
+inline __host__ __device__ void operator+=(uint2& a, uint2 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -338,7 +338,7 @@ inline __host__ __device__ uint2 operator+(uint b, uint2 a)
 {
   return make_uint2(a.x + b, a.y + b);
 }
-inline __host__ __device__ void operator+=(uint2 &a, uint b)
+inline __host__ __device__ void operator+=(uint2& a, uint b)
 {
   a.x += b;
   a.y += b;
@@ -348,7 +348,7 @@ inline __host__ __device__ float3 operator+(float3 a, float3 b)
 {
   return make_float3(a.x + b.x, a.y + b.y, a.z + b.z);
 }
-inline __host__ __device__ void operator+=(float3 &a, float3 b)
+inline __host__ __device__ void operator+=(float3& a, float3 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -358,7 +358,7 @@ inline __host__ __device__ float3 operator+(float3 a, float b)
 {
   return make_float3(a.x + b, a.y + b, a.z + b);
 }
-inline __host__ __device__ void operator+=(float3 &a, float b)
+inline __host__ __device__ void operator+=(float3& a, float b)
 {
   a.x += b;
   a.y += b;
@@ -369,7 +369,7 @@ inline __host__ __device__ int3 operator+(int3 a, int3 b)
 {
   return make_int3(a.x + b.x, a.y + b.y, a.z + b.z);
 }
-inline __host__ __device__ void operator+=(int3 &a, int3 b)
+inline __host__ __device__ void operator+=(int3& a, int3 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -379,7 +379,7 @@ inline __host__ __device__ int3 operator+(int3 a, int b)
 {
   return make_int3(a.x + b, a.y + b, a.z + b);
 }
-inline __host__ __device__ void operator+=(int3 &a, int b)
+inline __host__ __device__ void operator+=(int3& a, int b)
 {
   a.x += b;
   a.y += b;
@@ -390,7 +390,7 @@ inline __host__ __device__ uint3 operator+(uint3 a, uint3 b)
 {
   return make_uint3(a.x + b.x, a.y + b.y, a.z + b.z);
 }
-inline __host__ __device__ void operator+=(uint3 &a, uint3 b)
+inline __host__ __device__ void operator+=(uint3& a, uint3 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -400,7 +400,7 @@ inline __host__ __device__ uint3 operator+(uint3 a, uint b)
 {
   return make_uint3(a.x + b, a.y + b, a.z + b);
 }
-inline __host__ __device__ void operator+=(uint3 &a, uint b)
+inline __host__ __device__ void operator+=(uint3& a, uint b)
 {
   a.x += b;
   a.y += b;
@@ -424,7 +424,7 @@ inline __host__ __device__ float4 operator+(float4 a, float4 b)
 {
   return make_float4(a.x + b.x, a.y + b.y, a.z + b.z, a.w + b.w);
 }
-inline __host__ __device__ void operator+=(float4 &a, float4 b)
+inline __host__ __device__ void operator+=(float4& a, float4 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -439,7 +439,7 @@ inline __host__ __device__ float4 operator+(float b, float4 a)
 {
   return make_float4(a.x + b, a.y + b, a.z + b, a.w + b);
 }
-inline __host__ __device__ void operator+=(float4 &a, float b)
+inline __host__ __device__ void operator+=(float4& a, float b)
 {
   a.x += b;
   a.y += b;
@@ -451,7 +451,7 @@ inline __host__ __device__ int4 operator+(int4 a, int4 b)
 {
   return make_int4(a.x + b.x, a.y + b.y, a.z + b.z, a.w + b.w);
 }
-inline __host__ __device__ void operator+=(int4 &a, int4 b)
+inline __host__ __device__ void operator+=(int4& a, int4 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -466,7 +466,7 @@ inline __host__ __device__ int4 operator+(int b, int4 a)
 {
   return make_int4(a.x + b, a.y + b, a.z + b, a.w + b);
 }
-inline __host__ __device__ void operator+=(int4 &a, int b)
+inline __host__ __device__ void operator+=(int4& a, int b)
 {
   a.x += b;
   a.y += b;
@@ -478,7 +478,7 @@ inline __host__ __device__ uint4 operator+(uint4 a, uint4 b)
 {
   return make_uint4(a.x + b.x, a.y + b.y, a.z + b.z, a.w + b.w);
 }
-inline __host__ __device__ void operator+=(uint4 &a, uint4 b)
+inline __host__ __device__ void operator+=(uint4& a, uint4 b)
 {
   a.x += b.x;
   a.y += b.y;
@@ -493,7 +493,7 @@ inline __host__ __device__ uint4 operator+(uint b, uint4 a)
 {
   return make_uint4(a.x + b, a.y + b, a.z + b, a.w + b);
 }
-inline __host__ __device__ void operator+=(uint4 &a, uint b)
+inline __host__ __device__ void operator+=(uint4& a, uint b)
 {
   a.x += b;
   a.y += b;
@@ -509,7 +509,7 @@ inline __host__ __device__ float2 operator-(float2 a, float2 b)
 {
   return make_float2(a.x - b.x, a.y - b.y);
 }
-inline __host__ __device__ void operator-=(float2 &a, float2 b)
+inline __host__ __device__ void operator-=(float2& a, float2 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -522,7 +522,7 @@ inline __host__ __device__ float2 operator-(float b, float2 a)
 {
   return make_float2(b - a.x, b - a.y);
 }
-inline __host__ __device__ void operator-=(float2 &a, float b)
+inline __host__ __device__ void operator-=(float2& a, float b)
 {
   a.x -= b;
   a.y -= b;
@@ -532,7 +532,7 @@ inline __host__ __device__ int2 operator-(int2 a, int2 b)
 {
   return make_int2(a.x - b.x, a.y - b.y);
 }
-inline __host__ __device__ void operator-=(int2 &a, int2 b)
+inline __host__ __device__ void operator-=(int2& a, int2 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -545,7 +545,7 @@ inline __host__ __device__ int2 operator-(int b, int2 a)
 {
   return make_int2(b - a.x, b - a.y);
 }
-inline __host__ __device__ void operator-=(int2 &a, int b)
+inline __host__ __device__ void operator-=(int2& a, int b)
 {
   a.x -= b;
   a.y -= b;
@@ -555,7 +555,7 @@ inline __host__ __device__ uint2 operator-(uint2 a, uint2 b)
 {
   return make_uint2(a.x - b.x, a.y - b.y);
 }
-inline __host__ __device__ void operator-=(uint2 &a, uint2 b)
+inline __host__ __device__ void operator-=(uint2& a, uint2 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -568,7 +568,7 @@ inline __host__ __device__ uint2 operator-(uint b, uint2 a)
 {
   return make_uint2(b - a.x, b - a.y);
 }
-inline __host__ __device__ void operator-=(uint2 &a, uint b)
+inline __host__ __device__ void operator-=(uint2& a, uint b)
 {
   a.x -= b;
   a.y -= b;
@@ -578,7 +578,7 @@ inline __host__ __device__ float3 operator-(float3 a, float3 b)
 {
   return make_float3(a.x - b.x, a.y - b.y, a.z - b.z);
 }
-inline __host__ __device__ void operator-=(float3 &a, float3 b)
+inline __host__ __device__ void operator-=(float3& a, float3 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -592,7 +592,7 @@ inline __host__ __device__ float3 operator-(float b, float3 a)
 {
   return make_float3(b - a.x, b - a.y, b - a.z);
 }
-inline __host__ __device__ void operator-=(float3 &a, float b)
+inline __host__ __device__ void operator-=(float3& a, float b)
 {
   a.x -= b;
   a.y -= b;
@@ -603,7 +603,7 @@ inline __host__ __device__ int3 operator-(int3 a, int3 b)
 {
   return make_int3(a.x - b.x, a.y - b.y, a.z - b.z);
 }
-inline __host__ __device__ void operator-=(int3 &a, int3 b)
+inline __host__ __device__ void operator-=(int3& a, int3 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -617,7 +617,7 @@ inline __host__ __device__ int3 operator-(int b, int3 a)
 {
   return make_int3(b - a.x, b - a.y, b - a.z);
 }
-inline __host__ __device__ void operator-=(int3 &a, int b)
+inline __host__ __device__ void operator-=(int3& a, int b)
 {
   a.x -= b;
   a.y -= b;
@@ -628,7 +628,7 @@ inline __host__ __device__ uint3 operator-(uint3 a, uint3 b)
 {
   return make_uint3(a.x - b.x, a.y - b.y, a.z - b.z);
 }
-inline __host__ __device__ void operator-=(uint3 &a, uint3 b)
+inline __host__ __device__ void operator-=(uint3& a, uint3 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -642,7 +642,7 @@ inline __host__ __device__ uint3 operator-(uint b, uint3 a)
 {
   return make_uint3(b - a.x, b - a.y, b - a.z);
 }
-inline __host__ __device__ void operator-=(uint3 &a, uint b)
+inline __host__ __device__ void operator-=(uint3& a, uint b)
 {
   a.x -= b;
   a.y -= b;
@@ -653,7 +653,7 @@ inline __host__ __device__ float4 operator-(float4 a, float4 b)
 {
   return make_float4(a.x - b.x, a.y - b.y, a.z - b.z, a.w - b.w);
 }
-inline __host__ __device__ void operator-=(float4 &a, float4 b)
+inline __host__ __device__ void operator-=(float4& a, float4 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -664,7 +664,7 @@ inline __host__ __device__ float4 operator-(float4 a, float b)
 {
   return make_float4(a.x - b, a.y - b, a.z - b, a.w - b);
 }
-inline __host__ __device__ void operator-=(float4 &a, float b)
+inline __host__ __device__ void operator-=(float4& a, float b)
 {
   a.x -= b;
   a.y -= b;
@@ -676,7 +676,7 @@ inline __host__ __device__ int4 operator-(int4 a, int4 b)
 {
   return make_int4(a.x - b.x, a.y - b.y, a.z - b.z, a.w - b.w);
 }
-inline __host__ __device__ void operator-=(int4 &a, int4 b)
+inline __host__ __device__ void operator-=(int4& a, int4 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -691,7 +691,7 @@ inline __host__ __device__ int4 operator-(int b, int4 a)
 {
   return make_int4(b - a.x, b - a.y, b - a.z, b - a.w);
 }
-inline __host__ __device__ void operator-=(int4 &a, int b)
+inline __host__ __device__ void operator-=(int4& a, int b)
 {
   a.x -= b;
   a.y -= b;
@@ -703,7 +703,7 @@ inline __host__ __device__ uint4 operator-(uint4 a, uint4 b)
 {
   return make_uint4(a.x - b.x, a.y - b.y, a.z - b.z, a.w - b.w);
 }
-inline __host__ __device__ void operator-=(uint4 &a, uint4 b)
+inline __host__ __device__ void operator-=(uint4& a, uint4 b)
 {
   a.x -= b.x;
   a.y -= b.y;
@@ -718,7 +718,7 @@ inline __host__ __device__ uint4 operator-(uint b, uint4 a)
 {
   return make_uint4(b - a.x, b - a.y, b - a.z, b - a.w);
 }
-inline __host__ __device__ void operator-=(uint4 &a, uint b)
+inline __host__ __device__ void operator-=(uint4& a, uint b)
 {
   a.x -= b;
   a.y -= b;
@@ -734,7 +734,7 @@ inline __host__ __device__ float2 operator*(float2 a, float2 b)
 {
   return make_float2(a.x * b.x, a.y * b.y);
 }
-inline __host__ __device__ void operator*=(float2 &a, float2 b)
+inline __host__ __device__ void operator*=(float2& a, float2 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -747,7 +747,7 @@ inline __host__ __device__ float2 operator*(float b, float2 a)
 {
   return make_float2(b * a.x, b * a.y);
 }
-inline __host__ __device__ void operator*=(float2 &a, float b)
+inline __host__ __device__ void operator*=(float2& a, float b)
 {
   a.x *= b;
   a.y *= b;
@@ -757,7 +757,7 @@ inline __host__ __device__ int2 operator*(int2 a, int2 b)
 {
   return make_int2(a.x * b.x, a.y * b.y);
 }
-inline __host__ __device__ void operator*=(int2 &a, int2 b)
+inline __host__ __device__ void operator*=(int2& a, int2 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -770,7 +770,7 @@ inline __host__ __device__ int2 operator*(int b, int2 a)
 {
   return make_int2(b * a.x, b * a.y);
 }
-inline __host__ __device__ void operator*=(int2 &a, int b)
+inline __host__ __device__ void operator*=(int2& a, int b)
 {
   a.x *= b;
   a.y *= b;
@@ -780,7 +780,7 @@ inline __host__ __device__ uint2 operator*(uint2 a, uint2 b)
 {
   return make_uint2(a.x * b.x, a.y * b.y);
 }
-inline __host__ __device__ void operator*=(uint2 &a, uint2 b)
+inline __host__ __device__ void operator*=(uint2& a, uint2 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -793,7 +793,7 @@ inline __host__ __device__ uint2 operator*(uint b, uint2 a)
 {
   return make_uint2(b * a.x, b * a.y);
 }
-inline __host__ __device__ void operator*=(uint2 &a, uint b)
+inline __host__ __device__ void operator*=(uint2& a, uint b)
 {
   a.x *= b;
   a.y *= b;
@@ -803,7 +803,7 @@ inline __host__ __device__ float3 operator*(float3 a, float3 b)
 {
   return make_float3(a.x * b.x, a.y * b.y, a.z * b.z);
 }
-inline __host__ __device__ void operator*=(float3 &a, float3 b)
+inline __host__ __device__ void operator*=(float3& a, float3 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -817,7 +817,7 @@ inline __host__ __device__ float3 operator*(float b, float3 a)
 {
   return make_float3(b * a.x, b * a.y, b * a.z);
 }
-inline __host__ __device__ void operator*=(float3 &a, float b)
+inline __host__ __device__ void operator*=(float3& a, float b)
 {
   a.x *= b;
   a.y *= b;
@@ -828,7 +828,7 @@ inline __host__ __device__ int3 operator*(int3 a, int3 b)
 {
   return make_int3(a.x * b.x, a.y * b.y, a.z * b.z);
 }
-inline __host__ __device__ void operator*=(int3 &a, int3 b)
+inline __host__ __device__ void operator*=(int3& a, int3 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -842,7 +842,7 @@ inline __host__ __device__ int3 operator*(int b, int3 a)
 {
   return make_int3(b * a.x, b * a.y, b * a.z);
 }
-inline __host__ __device__ void operator*=(int3 &a, int b)
+inline __host__ __device__ void operator*=(int3& a, int b)
 {
   a.x *= b;
   a.y *= b;
@@ -853,7 +853,7 @@ inline __host__ __device__ uint3 operator*(uint3 a, uint3 b)
 {
   return make_uint3(a.x * b.x, a.y * b.y, a.z * b.z);
 }
-inline __host__ __device__ void operator*=(uint3 &a, uint3 b)
+inline __host__ __device__ void operator*=(uint3& a, uint3 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -867,7 +867,7 @@ inline __host__ __device__ uint3 operator*(uint b, uint3 a)
 {
   return make_uint3(b * a.x, b * a.y, b * a.z);
 }
-inline __host__ __device__ void operator*=(uint3 &a, uint b)
+inline __host__ __device__ void operator*=(uint3& a, uint b)
 {
   a.x *= b;
   a.y *= b;
@@ -878,7 +878,7 @@ inline __host__ __device__ float4 operator*(float4 a, float4 b)
 {
   return make_float4(a.x * b.x, a.y * b.y, a.z * b.z, a.w * b.w);
 }
-inline __host__ __device__ void operator*=(float4 &a, float4 b)
+inline __host__ __device__ void operator*=(float4& a, float4 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -893,7 +893,7 @@ inline __host__ __device__ float4 operator*(float b, float4 a)
 {
   return make_float4(b * a.x, b * a.y, b * a.z, b * a.w);
 }
-inline __host__ __device__ void operator*=(float4 &a, float b)
+inline __host__ __device__ void operator*=(float4& a, float b)
 {
   a.x *= b;
   a.y *= b;
@@ -905,7 +905,7 @@ inline __host__ __device__ int4 operator*(int4 a, int4 b)
 {
   return make_int4(a.x * b.x, a.y * b.y, a.z * b.z, a.w * b.w);
 }
-inline __host__ __device__ void operator*=(int4 &a, int4 b)
+inline __host__ __device__ void operator*=(int4& a, int4 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -920,7 +920,7 @@ inline __host__ __device__ int4 operator*(int b, int4 a)
 {
   return make_int4(b * a.x, b * a.y, b * a.z, b * a.w);
 }
-inline __host__ __device__ void operator*=(int4 &a, int b)
+inline __host__ __device__ void operator*=(int4& a, int b)
 {
   a.x *= b;
   a.y *= b;
@@ -932,7 +932,7 @@ inline __host__ __device__ uint4 operator*(uint4 a, uint4 b)
 {
   return make_uint4(a.x * b.x, a.y * b.y, a.z * b.z, a.w * b.w);
 }
-inline __host__ __device__ void operator*=(uint4 &a, uint4 b)
+inline __host__ __device__ void operator*=(uint4& a, uint4 b)
 {
   a.x *= b.x;
   a.y *= b.y;
@@ -947,7 +947,7 @@ inline __host__ __device__ uint4 operator*(uint b, uint4 a)
 {
   return make_uint4(b * a.x, b * a.y, b * a.z, b * a.w);
 }
-inline __host__ __device__ void operator*=(uint4 &a, uint b)
+inline __host__ __device__ void operator*=(uint4& a, uint b)
 {
   a.x *= b;
   a.y *= b;
@@ -963,7 +963,7 @@ inline __host__ __device__ float2 operator/(float2 a, float2 b)
 {
   return make_float2(a.x / b.x, a.y / b.y);
 }
-inline __host__ __device__ void operator/=(float2 &a, float2 b)
+inline __host__ __device__ void operator/=(float2& a, float2 b)
 {
   a.x /= b.x;
   a.y /= b.y;
@@ -972,7 +972,7 @@ inline __host__ __device__ float2 operator/(float2 a, float b)
 {
   return make_float2(a.x / b, a.y / b);
 }
-inline __host__ __device__ void operator/=(float2 &a, float b)
+inline __host__ __device__ void operator/=(float2& a, float b)
 {
   a.x /= b;
   a.y /= b;
@@ -986,7 +986,7 @@ inline __host__ __device__ float3 operator/(float3 a, float3 b)
 {
   return make_float3(a.x / b.x, a.y / b.y, a.z / b.z);
 }
-inline __host__ __device__ void operator/=(float3 &a, float3 b)
+inline __host__ __device__ void operator/=(float3& a, float3 b)
 {
   a.x /= b.x;
   a.y /= b.y;
@@ -996,7 +996,7 @@ inline __host__ __device__ float3 operator/(float3 a, float b)
 {
   return make_float3(a.x / b, a.y / b, a.z / b);
 }
-inline __host__ __device__ void operator/=(float3 &a, float b)
+inline __host__ __device__ void operator/=(float3& a, float b)
 {
   a.x /= b;
   a.y /= b;
@@ -1011,7 +1011,7 @@ inline __host__ __device__ float4 operator/(float4 a, float4 b)
 {
   return make_float4(a.x / b.x, a.y / b.y, a.z / b.z, a.w / b.w);
 }
-inline __host__ __device__ void operator/=(float4 &a, float4 b)
+inline __host__ __device__ void operator/=(float4& a, float4 b)
 {
   a.x /= b.x;
   a.y /= b.y;
@@ -1022,7 +1022,7 @@ inline __host__ __device__ float4 operator/(float4 a, float b)
 {
   return make_float4(a.x / b, a.y / b, a.z / b, a.w / b);
 }
-inline __host__ __device__ void operator/=(float4 &a, float b)
+inline __host__ __device__ void operator/=(float4& a, float b)
 {
   a.x /= b;
   a.y /= b;
