@@ -110,7 +110,7 @@ void Maps::perlin3D()
 
   PerlinNoise noise(info.seed);
 
-  std::vector<double> *v = new std::vector<double>;
+  std::vector<double>* v = new std::vector<double>;
   v->reserve(info.cloud->width);
   for (int i = 0; i < info.sizeX; ++i)
   {
@@ -164,56 +164,58 @@ void Maps::perlin3D()
   pcl2ros();
 }
 
-void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &maze)
+void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi& maze)
 {
   ROS_INFO("generating maze with width %d , height %d", xh - xl + 1, yh - yl + 1);
 
   if (xl < xh - 3 && yl < yh - 3)
-  { // the remaining area is larger than or equal to 5*5, need to add both x
+  {  // the remaining area is larger than or equal to 5*5, need to add both x
     // wall and y wall
-    bool valid = false; // used to judge whether the wall selection is valid
+    bool valid = false;  // used to judge whether the wall selection is valid
     int xm = 0;
     int ym = 0;
     ROS_INFO("entered 5*5 mode");
     while (valid == false)
     {
-      xm = (std::rand() % (xh - xl - 1) + xl + 1); // generating random number between xl+1 and xh-1(pointless to add a wall at the sides)
-      ym = (std::rand() % (yh - yl - 1) + yl + 1); // generating random number between yl+1 and yh-1(pointless to add a wall at the sides)
+      xm = (std::rand() % (xh - xl - 1) + xl + 1);  // generating random number between xl+1 and xh-1(pointless to add a
+                                                    // wall at the sides)
+      ym = (std::rand() % (yh - yl - 1) + yl + 1);  // generating random number between yl+1 and yh-1(pointless to add a
+                                                    // wall at the sides)
       if (xl - 1 >= 0)
-      { // there is a point at xl-1,ym
+      {  // there is a point at xl-1,ym
         if (maze(xl - 1, ym) == 0)
-        { // this is an opening,need to change random number
+        {  // this is an opening,need to change random number
           continue;
         }
       }
 
       else if (xh + 1 <= maze.cols() - 1)
-      { // there is a point at xh+1,ym
+      {  // there is a point at xh+1,ym
         if (maze(xh + 1, ym) == 0)
-        { // this is an opening,need to change random number
+        {  // this is an opening,need to change random number
           continue;
         }
       }
 
       else if (yl - 1 >= 0)
-      { // there is a point at xm,yl-1
+      {  // there is a point at xm,yl-1
         if (maze(xm, yl - 1) == 0)
-        { // this is an opening,need to change random number
+        {  // this is an opening,need to change random number
           continue;
         }
       }
 
       else if (yh + 1 <= maze.rows() - 1)
-      { // there is a point at xm,yh+1
+      {  // there is a point at xm,yh+1
         if (maze(xm, yh + 1) == 0)
-        { // this is an opening,need to change random number
+        {  // this is an opening,need to change random number
           continue;
         }
       }
 
       valid = true;
 
-    } // xm and ym are now the valid coordinate of the center of the wall
+    }  // xm and ym are now the valid coordinate of the center of the wall
     for (int i = xl; i <= xh; i++)
     {
       maze(i, ym) = 1;
@@ -221,39 +223,39 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
     for (int j = yl; j <= yh; j++)
     {
       maze(xm, j) = 1;
-    } // adding walls around the center point
+    }  // adding walls around the center point
     int d1 = std::rand() % (xm - xl) + xl;
     int d2 = std::rand() % (xh - xm) + xm + 1;
     int d3 = std::rand() % (ym - yl) + yl;
-    int d4 = std::rand() % (yh - ym) + ym + 1; // generating four possible door points
+    int d4 = std::rand() % (yh - ym) + ym + 1;  // generating four possible door points
 
-    int decision = std::rand() % 4; // random selection of three doors
+    int decision = std::rand() % 4;  // random selection of three doors
     switch (decision)
     {
-    case 0:
-      maze(d1, ym) = 0;
-      maze(d2, ym) = 0;
-      maze(xm, d3) = 0;
-      break;
+      case 0:
+        maze(d1, ym) = 0;
+        maze(d2, ym) = 0;
+        maze(xm, d3) = 0;
+        break;
 
-    case 1:
-      maze(d1, ym) = 0;
-      maze(d2, ym) = 0;
-      maze(xm, d4) = 0;
-      break;
+      case 1:
+        maze(d1, ym) = 0;
+        maze(d2, ym) = 0;
+        maze(xm, d4) = 0;
+        break;
 
-    case 2:
-      maze(d2, ym) = 0;
-      maze(xm, d3) = 0;
-      maze(xm, d4) = 0;
-      break;
+      case 2:
+        maze(d2, ym) = 0;
+        maze(xm, d3) = 0;
+        maze(xm, d4) = 0;
+        break;
 
-    case 3:
-      maze(d1, ym) = 0;
-      maze(xm, d3) = 0;
-      maze(xm, d4) = 0;
-      break;
-    } // the doors are opened for this cell
+      case 3:
+        maze(d1, ym) = 0;
+        maze(xm, d3) = 0;
+        maze(xm, d4) = 0;
+        break;
+    }  // the doors are opened for this cell
     if (yl - 1 >= 0)
     {
       if (maze(xm, yl - 1) == 0)
@@ -295,7 +297,7 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
     ROS_INFO("finished generating maze with width %d , height %d", xh - xl + 1, yh - yl + 1);
     std::cout << maze << std::endl;
     return;
-  } // when the remaining area is larger than or equal to 5*5
+  }  // when the remaining area is larger than or equal to 5*5
 
   else if (xl < xh - 2 && yl < yh - 2)
   {
@@ -303,9 +305,11 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
     int xm = 0;
     int ym = 0;
     int doorcount = 0;
-    xm = (std::rand() % (xh - xl - 1) + xl + 1); // generating random number between xl+1 and xh-1(pointless to add a wall at the sides)
-    ym = (std::rand() % (yh - yl - 1) + yl + 1); // generating random number between yl+1 and yh-1(pointless to add a wall at the sides)
-                                                 // xm and ym are now the valid coordinate of the center of the wall
+    xm = (std::rand() % (xh - xl - 1) + xl + 1);  // generating random number between xl+1 and xh-1(pointless to add a
+                                                  // wall at the sides)
+    ym = (std::rand() % (yh - yl - 1) + yl + 1);  // generating random number between yl+1 and yh-1(pointless to add a
+                                                  // wall at the sides) xm and ym are now the valid coordinate of the
+                                                  // center of the wall
     for (int i = xl; i <= xh; i++)
     {
       maze(i, ym) = 1;
@@ -313,7 +317,7 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
     for (int j = yl; j <= yh; j++)
     {
       maze(xm, j) = 1;
-    } // adding walls around the center point
+    }  // adding walls around the center point
     if (yl - 1 >= 0)
     {
       if (maze(xm, yl - 1) == 0)
@@ -353,35 +357,35 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
     int d1 = std::rand() % (xm - xl) + xl;
     int d2 = std::rand() % (xh - xm) + xm + 1;
     int d3 = std::rand() % (ym - yl) + yl;
-    int d4 = std::rand() % (yh - ym) + ym + 1; // generating four possible door points
+    int d4 = std::rand() % (yh - ym) + ym + 1;  // generating four possible door points
 
-    int decision = std::rand() % 4; // random selection of three doors
+    int decision = std::rand() % 4;  // random selection of three doors
     switch (decision)
     {
-    case 0:
-      maze(d1, ym) = 0;
-      maze(d2, ym) = 0;
-      maze(xm, d3) = 0;
-      break;
+      case 0:
+        maze(d1, ym) = 0;
+        maze(d2, ym) = 0;
+        maze(xm, d3) = 0;
+        break;
 
-    case 1:
-      maze(d1, ym) = 0;
-      maze(d2, ym) = 0;
-      maze(xm, d4) = 0;
-      break;
+      case 1:
+        maze(d1, ym) = 0;
+        maze(d2, ym) = 0;
+        maze(xm, d4) = 0;
+        break;
 
-    case 2:
-      maze(d2, ym) = 0;
-      maze(xm, d3) = 0;
-      maze(xm, d4) = 0;
-      break;
+      case 2:
+        maze(d2, ym) = 0;
+        maze(xm, d3) = 0;
+        maze(xm, d4) = 0;
+        break;
 
-    case 3:
-      maze(d1, ym) = 0;
-      maze(xm, d3) = 0;
-      maze(xm, d4) = 0;
-      break;
-    } // the doors are opened for this cell
+      case 3:
+        maze(d1, ym) = 0;
+        maze(xm, d3) = 0;
+        maze(xm, d4) = 0;
+        break;
+    }  // the doors are opened for this cell
     std::cout << maze << std::endl;
 
     ROS_INFO("finished generating maze with width %d , height %d", xh - xl + 1, yh - yl + 1);
@@ -390,14 +394,14 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
   }
 
   else if (xl < xh - 1 && yl < yh - 2)
-  { // the case of 3*4+
+  {  // the case of 3*4+
     ROS_INFO("entered 3*4+ mode");
     int doorcount = 0;
     int ym = 0;
     for (int i = yl; i <= yh; i++)
     {
       maze(xl + 1, i) = 1;
-    } // filling a center wall
+    }  // filling a center wall
     if (yl - 1 >= 0)
     {
       if (maze(xl + 1, yl - 1) == 0)
@@ -413,23 +417,23 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
         maze(xl + 1, yh) = 0;
         doorcount++;
       }
-    } // opening doors if the wall blocks the old doors
+    }  // opening doors if the wall blocks the old doors
     if (doorcount == 0)
     {
       ym = std::rand() % (yh - yl + 1) + yl;
       maze(xl + 1, ym) = 0;
     }
-  } // the case of 4+*3
+  }  // the case of 4+*3
   //
   else if (xl < xh - 2 && yl < yh - 1)
-  { // the case of 4+*3
+  {  // the case of 4+*3
     ROS_INFO("entered 4+*3 mode");
     int doorcount = 0;
     int xm = 0;
     for (int i = xl; i <= xh; i++)
     {
       maze(i, yl + 1) = 1;
-    } // filling a center wall
+    }  // filling a center wall
     if (xl - 1 >= 0)
     {
       if (maze(xl - 1, yl + 1) == 0)
@@ -445,16 +449,16 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
         maze(xh, yl + 1) = 0;
         doorcount++;
       }
-    } // opening doors if the wall blocks the old doors
+    }  // opening doors if the wall blocks the old doors
     if (doorcount == 0)
     {
       xm = std::rand() % (xh - xl + 1) + xl;
       maze(xm, yl + 1) = 0;
     }
-  } // the case of 4+*3
+  }  // the case of 4+*3
 
   else if (xl < xh - 1 && yl < yh - 1)
-  { // the case of 3*3
+  {  // the case of 3*3
     maze(xl + 1, yl + 1) = 1;
     return;
   }
@@ -465,7 +469,7 @@ void Maps::recursiveDivision(int xl, int xh, int yl, int yh, Eigen::MatrixXi &ma
   }
 }
 
-void Maps::recursizeDivisionMaze(Eigen::MatrixXi &maze)
+void Maps::recursizeDivisionMaze(Eigen::MatrixXi& maze)
 {
   //! @todo all bugs here...
   int sx = maze.rows();
@@ -526,18 +530,18 @@ void Maps::recursizeDivisionMaze(Eigen::MatrixXi &maze)
   }
   switch (std::rand() % 4)
   {
-  case 0:
-    maze(x1, py) = 1;
-    break;
-  case 1:
-    maze(x2, py) = 1;
-    break;
-  case 2:
-    maze(px, y1) = 1;
-    break;
-  case 3:
-    maze(px, y2) = 1;
-    break;
+    case 0:
+      maze(x1, py) = 1;
+      break;
+    case 1:
+      maze(x2, py) = 1;
+      break;
+    case 2:
+      maze(px, y1) = 1;
+      break;
+    case 3:
+      maze(px, y2) = 1;
+      break;
   }
 
   if (px > 2 && py > 2)
@@ -560,7 +564,6 @@ void Maps::recursizeDivisionMaze(Eigen::MatrixXi &maze)
   }
   if ((sx - px - 1) > 2 && (sy - py - 1) > 2)
   {
-
     Eigen::MatrixXi sub = maze.block(px, py, sy - px, sy - py);
 
     recursizeDivisionMaze(sub);
@@ -587,9 +590,9 @@ void Maps::maze2D()
 
   switch (type)
   {
-  case 1:
-    recursiveDivision(0, maze.cols() - 1, 0, maze.rows() - 1, maze);
-    break;
+    case 1:
+      recursiveDivision(0, maze.cols() - 1, 0, maze.rows() - 1, maze);
+      break;
   }
 
   if (addWallX)
@@ -624,10 +627,8 @@ void Maps::maze2D()
             for (int k = 0; k < info.sizeZ; ++k)
             {
               pcl::PointXYZ pt_random;
-              pt_random.x =
-                  i * width + ii / info.scale - info.sizeX / (2.0 * info.scale);
-              pt_random.y =
-                  j * width + jj / info.scale - info.sizeY / (2.0 * info.scale);
+              pt_random.x = i * width + ii / info.scale - info.sizeX / (2.0 * info.scale);
+              pt_random.y = j * width + jj / info.scale - info.sizeY / (2.0 * info.scale);
               pt_random.z = k / info.scale;
               info.cloud->points.push_back(pt_random);
             }
@@ -647,7 +648,7 @@ Maps::BasicInfo Maps::getInfo() const
   return info;
 }
 
-void Maps::setInfo(const BasicInfo &value)
+void Maps::setInfo(const BasicInfo& value)
 {
   info = value;
 }
@@ -660,21 +661,21 @@ void Maps::generate(int type)
 {
   switch (type)
   {
-  default:
-  case 1:
-    perlin3D();
-    break;
-  case 2:
-    randomMapGenerate();
-    break;
-  case 3:
-    std::srand(info.seed);
-    maze2D();
-    break;
-  case 4: // generating 3d maze
-    std::srand(info.seed);
-    Maze3DGen();
-    break;
+    default:
+    case 1:
+      perlin3D();
+      break;
+    case 2:
+      randomMapGenerate();
+      break;
+    case 3:
+      std::srand(info.seed);
+      maze2D();
+      break;
+    case 4:  // generating 3d maze
+      std::srand(info.seed);
+      Maze3DGen();
+      break;
   }
 }
 
@@ -756,31 +757,29 @@ void Maps::Maze3DGen()
     pt_random.y = ry;
     pt_random.z = rz;
     base.push_back(pt_random);
-  } // generating random cores in the space
+  }  // generating random cores in the space
 
   for (int i = 0; i < info.sizeX; i++)
   {
     for (int j = 0; j < info.sizeY; j++)
     {
       for (int k = 0; k < info.sizeZ; k++)
-      { // for every scaled coordinate points
+      {  // for every scaled coordinate points
         pcl::PointXYZ test;
         test.x = i / info.scale - info.sizeX / (2 * info.scale);
         test.y = j / info.scale - info.sizeY / (2 * info.scale);
-        test.z = k / info.scale - info.sizeZ / (2 * info.scale); // marking the corresponding point location
+        test.z = k / info.scale - info.sizeZ / (2 * info.scale);  // marking the corresponding point location
 
         MazePoint mp;
         mp.setPoint(test);
         mp.setPoint2(-1);
         mp.setPoint1(-1);
         mp.setDist1(10000.0);
-        mp.setDist2(100000.0); // setting super large starting values
+        mp.setDist2(100000.0);  // setting super large starting values
         for (int ii = 0; ii < numNodes; ii++)
         {
-          double dist =
-              std::sqrt((base[ii].x - test.x) * (base[ii].x - test.x) +
-                        (base[ii].y - test.y) * (base[ii].y - test.y) +
-                        (base[ii].z - test.z) * (base[ii].z - test.z));
+          double dist = std::sqrt((base[ii].x - test.x) * (base[ii].x - test.x) + (base[ii].y - test.y) * (base[ii].y - test.y) +
+                                  (base[ii].z - test.z) * (base[ii].z - test.z));
           if (dist < mp.getDist1())
           {
             mp.setDist2(mp.getDist1());
@@ -793,20 +792,16 @@ void Maps::Maze3DGen()
           {
             mp.setDist2(dist);
             mp.setPoint2(ii);
-          } // finding the distances to the nearest two cores
+          }  // finding the distances to the nearest two cores
         }
         if (std::abs(mp.getDist2() - mp.getDist1()) < 1 / info.scale)
-        { // the tested location is on one of the middle planes
+        {  // the tested location is on one of the middle planes
           if ((mp.getPoint1() + mp.getPoint2()) > int((1 - connectivity) * numNodes) &&
               (mp.getPoint1() + mp.getPoint2()) < int((1 + connectivity) * numNodes))
-          { // this is a holed wall
-            double judge =
-                std::sqrt((base[mp.getPoint1()].x - base[mp.getPoint2()].x) *
-                              (base[mp.getPoint1()].x - base[mp.getPoint2()].x) +
-                          (base[mp.getPoint1()].y - base[mp.getPoint2()].y) *
-                              (base[mp.getPoint1()].y - base[mp.getPoint2()].y) +
-                          (base[mp.getPoint1()].z - base[mp.getPoint2()].z) *
-                              (base[mp.getPoint1()].z - base[mp.getPoint2()].z));
+          {  // this is a holed wall
+            double judge = std::sqrt((base[mp.getPoint1()].x - base[mp.getPoint2()].x) * (base[mp.getPoint1()].x - base[mp.getPoint2()].x) +
+                                     (base[mp.getPoint1()].y - base[mp.getPoint2()].y) * (base[mp.getPoint1()].y - base[mp.getPoint2()].y) +
+                                     (base[mp.getPoint1()].z - base[mp.getPoint2()].z) * (base[mp.getPoint1()].z - base[mp.getPoint2()].z));
             if (mp.getDist1() + mp.getDist2() - judge >= roadRad / (info.scale * 3))
             {
               info.cloud->points.push_back(mp.getPoint());
