@@ -10,18 +10,15 @@ public:
 
   void setMass(const double mass);
   void setGravity(const double g);
-  void setPosition(const Eigen::Vector3d &position);
-  void setVelocity(const Eigen::Vector3d &velocity);
-  void setAcc(const Eigen::Vector3d &acc);
+  void setPosition(const Eigen::Vector3d& position);
+  void setVelocity(const Eigen::Vector3d& velocity);
+  void setAcc(const Eigen::Vector3d& acc);
 
-  void calculateControl(const Eigen::Vector3d &des_pos,
-                        const Eigen::Vector3d &des_vel,
-                        const Eigen::Vector3d &des_acc, const double des_yaw,
-                        const double des_yaw_dot, const Eigen::Vector3d &kx,
-                        const Eigen::Vector3d &kv);
+  void calculateControl(const Eigen::Vector3d& des_pos, const Eigen::Vector3d& des_vel, const Eigen::Vector3d& des_acc, const double des_yaw,
+                        const double des_yaw_dot, const Eigen::Vector3d& kx, const Eigen::Vector3d& kv);
 
-  const Eigen::Vector3d &getComputedForce();
-  const Eigen::Quaterniond &getComputedOrientation();
+  const Eigen::Vector3d& getComputedForce();
+  const Eigen::Quaterniond& getComputedOrientation();
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
 private:
