@@ -9,9 +9,9 @@
 
 namespace quadrotor_msgs
 {
-  bool decodeOutputData(const std::vector<uint8_t> &data, quadrotor_msgs::OutputData &output);
-  bool decodeStatusData(const std::vector<uint8_t> &data, quadrotor_msgs::StatusData &status);
-  bool decodePPROutputData(const std::vector<uint8_t> &data, quadrotor_msgs::PPROutputData &output);
-}
+bool decodeOutputData(const std::vector<uint8_t>& data, quadrotor_msgs::OutputData& output);
+bool decodeStatusData(const std::vector<uint8_t>& data, quadrotor_msgs::StatusData& status);
+bool decodePPROutputData(const std::vector<uint8_t>& data, quadrotor_msgs::PPROutputData& output);
+}  // namespace quadrotor_msgs
 
 #endif
