@@ -31,11 +31,11 @@
 #define RVIZ_GOAL_TOOL_H
 
 #ifndef Q_MOC_RUN  // See: https://bugreports.qt-project.org/browse/QTBUG-22829
-# include <QObject>
+#include <QObject>
 
-# include <ros/ros.h>
+#include <ros/ros.h>
 
-# include "pose_tool.h"
+#include "pose_tool.h"
 #endif
 
 namespace rviz
@@ -44,12 +44,14 @@ class Arrow;
 class DisplayContext;
 class StringProperty;
 
-class Goal3DTool: public Pose3DTool
+class Goal3DTool : public Pose3DTool
 {
-Q_OBJECT
+  Q_OBJECT
 public:
   Goal3DTool();
-  virtual ~Goal3DTool() {}
+  virtual ~Goal3DTool()
+  {
+  }
   virtual void onInitialize();
 
 protected:
@@ -65,8 +67,6 @@ private:
   StringProperty* topic_property_;
 };
 
-}
+}  // namespace rviz
 
 #endif
-
-
