@@ -43,7 +43,7 @@ namespace rviz
 class Arrow;
 class DisplayContext;
 
-class Pose3DTool: public Tool
+class Pose3DTool : public Tool
 {
 public:
   Pose3DTool();
@@ -54,7 +54,7 @@ public:
   virtual void activate();
   virtual void deactivate();
 
-  virtual int processMouseEvent( ViewportMouseEvent& event );
+  virtual int processMouseEvent(ViewportMouseEvent& event);
 
 protected:
   virtual void onPoseSet(double x, double y, double z, double theta) = 0;
@@ -73,8 +73,6 @@ protected:
   Ogre::Vector3 pos_;
 };
 
-}
+}  // namespace rviz
 
 #endif
-
-
