@@ -12,37 +12,34 @@ from sensor_msgs.msg import Joy
 pub = None
 pub1 = None
 
-def callback(odom_msg):
-    q = np.array([odom_msg.pose.pose.orientation.x,
-                  odom_msg.pose.pose.orientation.y,
-                  odom_msg.pose.pose.orientation.z,
-                  odom_msg.pose.pose.orientation.w])
 
-    e = tfs.euler_from_quaternion(q, 'rzyx')
+def callback(odom_msg):
+    q = np.array([odom_msg.pose.pose.orientation.x, odom_msg.pose.pose.orientation.y, odom_msg.pose.pose.orientation.z, odom_msg.pose.pose.orientation.w])
+
+    e = tfs.euler_from_quaternion(q, "rzyx")
 
     euler_msg = Vector3Stamped()
     euler_msg.header = odom_msg.header
-    euler_msg.vector.z = e[0]*180.0/3.14159
-    euler_msg.vector.y = e[1]*180.0/3.14159
-    euler_msg.vector.x = e[2]*180.0/3.14159
+    euler_msg.vector.z = e[0] * 180.0 / 3.14159
+    euler_msg.vector.y = e[1] * 180.0 / 3.14159
+    euler_msg.vector.x = e[2] * 180.0 / 3.14159
 
     pub.publish(euler_msg)
 
-def imu_callback(imu_msg):
-    q = np.array([imu_msg.orientation.x,
-                  imu_msg.orientation.y,
-                  imu_msg.orientation.z,
-                  imu_msg.orientation.w])
 
-    e = tfs.euler_from_quaternion(q, 'rzyx')
+def imu_callback(imu_msg):
+    q = np.array([imu_msg.orientation.x, imu_msg.orientation.y, imu_msg.orientation.z, imu_msg.orientation.w])
+
+    e = tfs.euler_from_quaternion(q, "rzyx")
 
     euler_msg = Vector3Stamped()
     euler_msg.header = imu_msg.header
-    euler_msg.vector.z = e[0]*180.0/3.14159
-    euler_msg.vector.y = e[1]*180.0/3.14159
-    euler_msg.vector.x = e[2]*180.0/3.14159
+    euler_msg.vector.z = e[0] * 180.0 / 3.14159
+    euler_msg.vector.y = e[1] * 180.0 / 3.14159
+    euler_msg.vector.x = e[2] * 180.0 / 3.14159
 
     pub1.publish(euler_msg)
+
 
 def joy_callback(joy_msg):
     out_msg = Vector3Stamped()
